backend/requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary  # PostgreSQL 的 Python 驱动
pydantic-settings
celery[redis]    # 安装 Celery 及其 Redis 依赖
python-dotenv
openai-whisper # OpenAI Whisper 的开源实现
torch           # PyTorch 深度学习框架
torchvision     # PyTorch 的计算机视觉库
torchaudio      # PyTorch 的音频库
transformers    # Hugging Face 库，用于加载 CLIP
Pillow          # 图像处理库
chromadb        # ChromaDB 的 Python 客户端
sentence-transformers